Refactor viewModel and add manual alert trigger
